+++ src/components/EventDialog.tsx
@@ -7,7 +7,7 @@ import TextField from "@mui/material/TextField";
 import type { TSchema } from "@sinclair/typebox";
 import { TypeCompiler } from "@sinclair/typebox/compiler";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import { getISODateString } from "./utils";
+import { getISODateString } from "../shared/utils";
 import type { DialogButton } from "@/types/dialog";
 import type { Event } from "@/types/event";
